requirements.txt
pandas==2.2.2
numpy==1.26.4
opencv-python==4.9.0.80
tensorflow==2.16.1
pyaml==6.0.1
uvicorn==0.29.0
pillow==10.3.0